* gui-gtk/gtk-tile-set.c (scale_and_paint_svg_image): Change a lot of types and use explicit casting to suppress all GCC 4.0 warnings.
* gui-gtk/gtk-named-vbox.h (_GtkNamedVBox): Change type of
`left_padding' field to `int' to avoid GCC 4.0 warning.

* gui-gtk/gtk-gtp-client-interface.c (handle_engine_output):
Change type of `bytes_read' to `gsize' to silence GCC 4.0 warning.
(handle_engine_stderr): Likewise.

* sgf/sgf-utils.c (sgf_utils_descend_nodes): Explicitly cast an
argument of sgf_node_get_number_property_value() to `int *' to
prevent GCC 4.0 warning.

* utils/buffered-writer.c (buffered_writer_add_character): Change
type of `buffer_size' and some local variables to
`size_t'.  (Dealing with GCC 4.0 warnings.)
(buffered_writer_add_newline): Likewise for local variables.
(buffered_writer_cat_as_string): Likewise for `length' and local
variables.
(buffered_writer_cat_as_strings): Likewise for length arguments
from variadic argument list.
(buffered_writer_vprintf): Likewise for a local variable.
(buffered_writer_vcprintf): Likewise.
(update_column): Likewise for `length' and a local variable.


git-svn-id: svn://svn.gna.org/quarry/trunk@722 3f990595-2723-0410-941d-8edc91ad4238

+++ src/gui-gtk/gtk-gtp-client-interface.c
@@ -241,7 +241,7 @@ handle_engine_output (GIOChannel *read_output_from, GIOCondition condition,
 {
   if (!data->deletion_scheduled) {
     gchar buffer[0x1000];
-    gint bytes_read;
+    gsize bytes_read;
 
     /* Check that the channel is readable.  This loop throws
      * responses at the client and it can decide to delete itself
@@ -299,7 +299,7 @@ handle_engine_stderr (GIOChannel *read_output_from, GIOCondition condition,
 {
   if (!data->deletion_scheduled) {
     gchar buffer[0x1000];
-    gint bytes_read;
+    gsize bytes_read;
 
     while (g_io_channel_get_flags (read_output_from)
 	   & G_IO_FLAG_IS_READABLE) {
